requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.6
python-flutterwave==0.7.9
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
